Added simple setup of provider

+++ index.js
@@ -1,36 +1,30 @@
 #!/usr/bin/env node
 
-const prompts = require('prompts')
 const fs = require('fs')
+const https = require('https')
 const path = require('path')
 const childProcess = require('child_process')
 const chalk = require('chalk')
 const ora = require('ora');
 
+const fetch = url => (new Promise((resolve, reject) => {
+  https.get(url, response => {
+    if (response.statusCode === 302) {
+      return resolve(fetch(`https://${response.req._headers.host}${response.headers.location}`))
+    }
+    let data = ''
 
-const rawTarget = process.argv[2]
-
-if (!rawTarget) {
-  console.error('You must specify a target folder')
-  return 1
-}
-
-const targetFolder = path.resolve(rawTarget)
-
-if (!fs.existsSync(path.resolve(targetFolder, '..'))) {
-  console.error('The parent folder does not exist')
-  return 1
-}
-
-
-fs.mkdirSync(targetFolder, { recursive: true })
+    response.on('data', chunk => data += chunk)
+    response.on('end', () => resolve(JSON.parse(data)))
+  }).on('error', error => reject(`Error: ${err.message}`))
+}))
 
 const spawn = (command, arguments, options) =>
   childProcess.spawn(command, arguments, {shell: process.platform == 'win32', ...options})
 
-const run = command => new Promise((resolve, reject) => {
+const run = (command, directory) => new Promise((resolve, reject) => {
   const splitCommand = command.split(' ')
-  const childProcess = spawn(splitCommand[0], splitCommand.slice(1), {cwd: targetFolder})
+  const childProcess = spawn(splitCommand[0], splitCommand.slice(1), {cwd: directory || path.resolve('.')})
   childProcess.stderr.pipe(process.stderr)
   childProcess.on('close', code => {
     if (code === 0) {
@@ -41,10 +35,10 @@ const run = command => new Promise((resolve, reject) => {
   })
 })
 
-const install = package => () => new Promise((resolve, reject) => {
+const install = (package, directory) => () => new Promise((resolve, reject) => {
   const spinner = ora(`installing ${package}`).start()
 
-  const childProcess = spawn('npm', ['install', package, '--silent'], {cwd: targetFolder})
+  const childProcess = spawn('npm', ['install', package, '--silent'], {cwd: directory || path.resolve('.')})
   childProcess.stderr.pipe(process.stderr)
   childProcess.on('close', code => {
     if (code === 0) {
@@ -57,25 +51,65 @@ const install = package => () => new Promise((resolve, reject) => {
   })
 })
 
-run('npm init -y')
-  .then(() => {
-    const package = JSON.parse(fs.readFileSync(path.resolve(targetFolder, 'package.json'), 'utf-8'))
-    package.private = true
-    package.scripts = { start: 'node index.js' }
-    delete package.description
-    delete package.keywords
-    delete package.author
-    delete package.license
-    fs.writeFileSync(path.resolve(targetFolder, 'package.json'), JSON.stringify(package, null, 2))
-  })
-  .then(install('@kanshi/kanshi-sha'))
-  .then(install('@kanshi/kanshi'))
-  .then(() => {
-    fs.copyFileSync('setup/launcher.js', path.resolve(targetFolder, 'index.js'))
-    fs.mkdirSync(path.resolve(targetFolder, 'configuration'))
-    fs.copyFileSync('setup/standalones.js', path.resolve(targetFolder, 'configuration', 'standalones.js'))
-    fs.copyFileSync('setup/packages.js', path.resolve(targetFolder, 'configuration', 'packages.js'))
-    fs.copyFileSync('setup/providers.js', path.resolve(targetFolder, 'configuration', 'providers.js'))
-    fs.mkdirSync(path.resolve(targetFolder, 'providers'))
-  })
-  .catch(console.error)
+
+const argument = process.argv[2]
+
+if (!argument) {
+  console.error('You must specify a target folder')
+  return 1
+}
+
+if (argument !== 'install' || !fs.existsSync(path.resolve('.', 'package.json')) && !process.argv[3]) {
+  const targetFolder = path.resolve(argument)
+
+  if (!fs.existsSync(path.resolve(targetFolder, '..'))) {
+    console.error('The parent folder does not exist')
+    return 1
+  }
+
+
+  fs.mkdirSync(targetFolder, { recursive: true })
+
+  run('npm init -y', targetFolder)
+    .then(() => {
+      const package = JSON.parse(fs.readFileSync(path.resolve(targetFolder, 'package.json'), 'utf-8'))
+      package.private = true
+      package.scripts = { start: 'node index.js' }
+      delete package.description
+      delete package.keywords
+      delete package.author
+      delete package.license
+      fs.writeFileSync(path.resolve(targetFolder, 'package.json'), JSON.stringify(package, null, 2))
+    })
+    .then(install('@kanshi/kanshi-sha', targetFolder))
+    .then(install('@kanshi/kanshi', targetFolder))
+    .then(() => {
+      fs.copyFileSync('setup/launcher.js', path.resolve(targetFolder, 'index.js'))
+      fs.mkdirSync(path.resolve(targetFolder, 'configuration'))
+      fs.copyFileSync('setup/standalones.js', path.resolve(targetFolder, 'configuration', 'standalones.js'))
+      fs.copyFileSync('setup/packages.js', path.resolve(targetFolder, 'configuration', 'packages.js'))
+      fs.copyFileSync('setup/providers.js', path.resolve(targetFolder, 'configuration', 'providers.js'))
+      fs.mkdirSync(path.resolve(targetFolder, 'providers'))
+    })
+    .catch(console.error)
+} else if (argument === 'install') {
+  if (!process.argv[3]) {
+    console.error("You didn't specify what provider you wanted to install")
+    return 1
+  }
+
+  if (!fs.existsSync(path.resolve('.', 'package.json'))) {
+    console.error('Missing package.json file. Are you sure you are in the right folder?')
+    return 1
+  }
+
+  const provider = process.argv[3]
+
+  fetch(`https://unpkg.com/${provider}/kanshi.json`)
+    .then(install(provider))
+    .then(() => {
+      const manifest = JSON.parse(fs.readFileSync(path.resolve('node_modules', provider, 'kanshi.json'), 'utf-8'))
+      fs.writeFileSync(path.resolve('providers', `${manifest.name}.js`), `module.exports = require('${provider}')`)
+    })
+    .catch(console.error)
+}
